requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.24.30
botocore==1.27.30
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.8.0
django-robots==5.0
django-storages==1.12.3
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.14.0
idna==3.3
jmespath==1.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
Pillow==9.3.0
protobuf==4.21.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.2.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
